Fixed the padding of the ViewHolder fast run button. Set up the button. Fixed the parcelable in method in Program after changing ID to a Long

+++ app/src/main/java/com/tpb/brainfuck_app/Program.java
@@ -23,10 +23,8 @@ public class Program implements Parcelable {
 
     public Program() {}
 
-
-
     protected Program(Parcel in) {
-        id = in.readInt();
+        id = in.readLong();
         name = in.readString();
         desc = in.readString();
         prog = in.readString();
